fix : header height push
