+++ src/app/signup/userdata/style.css
@@ -52,7 +52,7 @@
   z-index: 2;
 }
 
-.wrapper {
+.scrollanimation {
   animation: scroll 5s 1;
   margin: 0;
   overflow: hidden;
